+++ nflow-engine/src/test/java/io/nflow/engine/service/MaintenanceConfigurationDeserializationTest.java
@@ -8,15 +8,13 @@ import static org.joda.time.Period.days;
 
 import org.junit.jupiter.api.Test;
 
-import tools.jackson.databind.ObjectMapper;
+import io.nflow.engine.service.MaintenanceConfiguration.ConfigurationItem;
 import tools.jackson.databind.json.JsonMapper;
 import tools.jackson.datatype.joda.JodaModule;
 
-import io.nflow.engine.service.MaintenanceConfiguration.ConfigurationItem;
-
 class MaintenanceConfigurationDeserializationTest {
 
-  private final ObjectMapper mapper;
+  private final JsonMapper mapper;
 
   MaintenanceConfigurationDeserializationTest() {
     mapper = JsonMapper.builder()