+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 **Highlights**
 
 - POTENTIALLY BREAKING CHANGE: `StateExecution.setVariable(...)` throws error when called for `@StateVar` state variables (see details below)
+- Upgrade to Spring 7
+- BREAKING CHANGE: Replace `ObjectMapper` with `JsonMapper`
+  - `EngineObjectMapperSupplier` renamed to `EngineJsonMapperSupplier`
+  - `EngineConfiguration.nflowObjectMapper()` renamed to `EngineConfiguration.nflowJsonMapper()`
+  - `EngineModule.nflowObjectMapper()` renamed to `EngineModule.nflowJsonMapper()`
+  - `RestConfiguration.nflowRestObjectMapper()` renamed to `RestConfiguration.nflowRestJsonMapper()`
+  - `RestConfiguration.REST_OBJECT_MAPPER` renamed to `RestConfiguration.REST_JSON_MAPPER`
+  - `RestConfiguration.objectMapper()` renamed to `RestConfiguration.jsonMapper()`
+  - `RestClientConfiguration.objectMapper()` renamed to `RestClientConfiguration.jsonMapper()`
 
 **Details**
 
@@ -27,7 +36,6 @@
 **Highlights**
 
 - Dependency updates
-- Upgrade to Spring 7
 
 **Details**